+++ src/main/java/com/cinchapi/runway/LoadRecordSelection.java
@@ -61,6 +61,11 @@ final class LoadRecordSelection<T extends Record> extends DatabaseSelection<T> {
     }
 
     @Override
+    Object $emptyResultConstant() {
+        return null;
+    }
+
+    @Override
     boolean isCombinable() {
         return true;
     }